Camera & A blue-to-white gradient

+++ RTXAppln/core/source/ImGuiLayer.cpp
@@ -68,7 +68,7 @@ void ImGuiLayer::OnRender ()
     }
     ImGui::End ();
 
-    ImGui::SetNextWindowSize(ImVec2(m_ViewportWidth, m_ViewportHeight));
+    ImGui::SetNextWindowSize(ImVec2(static_cast<float>(m_ViewportWidth), static_cast<float>(m_ViewportHeight)));
     ImGui::PushStyleVar(ImGuiStyleVar_WindowPadding, ImVec2(0.0f, 0.0f));
     ImGui::Begin("Viewport", nullptr,
         ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_NoScrollWithMouse);
